fix #10398 "Large values (> 2^32) can cause gas to fail"
Set a bit in fixS struct so that large value can still be
handled. Without it, assembler will fail in fixup_segment when it checks
there is enough bits to store the value. As we usually split our reloc
across different words, it may think that 4 bytes are not enough for
large value. This simply skips the tests.