K1: better support of __stacksize
__stacksize is used on Linux to choose the stack size of an application.
In case on non-MMU it also defines the amount of memory available by
brk(). The old implementation didn't allow an override of the option
by using -Wl,--defsym,__stacksize=<size of the stack> at compile time.
The new implementation is borrowed from blackfind and uses a generic
function doing exactly what we want.

Signed-off-by: Marta Rybczynska <marta.rybczynska@kalray.eu>
Signed-off-by: Marc Poulhies <mpoulhies@kalray.eu>