Force read registers inside a dummy call
After the change in gdbserver to read only a few registers when a
processors stops, the registers restoration called at the end of the
dummy call writes all modified or unknown registers one by one. This
is slow and ISS does not support the write of some registers. To
prevent this restauration cause by the fact that the registers are
not known, this commit reads all registers inside dummy calls

	*gdb/k1-common-tdep.c (k1_dummy_id): Read all registers