Support so libraries enumaration during the library list change
Ref T10284

Summary:
When the MPPA code loads/unloads a library, the .so libraries list should
not be used. If the gdb libraries enumeration happens during this process,
the previous list will be provided, if it is available. At the end of
the list change, gdb will be anyway notified and the enumaration of the new
list be performed.

	*gdb/solib-k1-bare.c: Add valid field to the mppa_dl_debug_s struct
	and valid_addr and last_head fields to the struct k1_bare_solib_info
	struct
	(k1_bare_solib_load_debug_info): Save the address of the valid field
	(k1_bare_solib_dup_chain): Duplicates a solib list to be available
	next time is requested and the solib list is not valid
	(k1_bare_solib_free_chain): Deletes a solib list
	(k1_bare_solib_current_sos): If the solib is not valid (is under
	change), provide the old list

Test Plan: Tested by the opencl testsuit

Reviewers: O12 gdb!

Differential Revision: https://phab.kalray.eu/D2383