Add support for in-place stepi over hardware breakpoints
Ref T11296

Summary:
When gdb wants to step over a hardware breakpoint on an unmapped address,
gdb tries to execute the bundle from another location (displace
execution). As the source memory is not mapped, the memory copy to the
displaced location will fail and gdb will try to execute bundle in-place.
Before doing this, it will wait for all threads of the current process (all
CPUs of the current cluster) to stop, which would very probably lead to gdb
stall because not all CPUs can be stopped (maybe they are not started).
To avoid this problem, we disable the hardware breakpoint only on the
current CPU (feature not supported by gdb), execute the stepi in-place
and re-enable the hardware breakpoint on that CPU.

	*gdb/k1-tdep.c (struct displaced_step_closure): Add the
	execute_inplace field which specifies if the current displace is
	executed in-place
	(k1_is_mmu_enabled): New function which returns 1 if the mmu is
	enabled for the thread given by the provided regs
	(k1_displaced_step_location): Use the new k1_is_mmu_enabled
	function
	(k1_enable_hbkp_on_cpu): New function to send to gdbstub the
	request of enabling/disabling the hardware breakpoint on the curent
	CPU
	(k1_displaced_step_copy_insn): If the memory cannot accessed, MMU
	is enabled and there is a hardware breakpoint at this location,
	activate the in-place execution
	(k1_displaced_step_fixup): Re-enable the hardware breakpoint after
	the stepi has been executed
	(k1_write_pc): New function that makes possible the in-place
	execution
	(k1_gdbarch_init): Set the gdbarch write pc function
	*gdb/k1-common-tdep.h: Move some Kalray register field bit position
	declarations from gdb/k1-exception-info.c
	*gdb/k1-exception-info.c: See above

Test Plan: Insert a hardware breakpoint at an unmmaped address and stepi

Reviewers: O12 gdb!

Differential Revision: https://phab.kalray.eu/D2873