Fix bad TLS relocation for static cases
When ld resolves tpoff relocation, it uses a TCB_SIZE for offsetting
the start address: $tp would point to TCB_SIZE bytes before the start
of the static TLS block. That's how it works on aarch64 from which our
code is based. Unfortunately, our $tp has NO offset and points after
the TCB. Moreover, our TCB is 8bytes and not 16 as the code states.

ref T11849

	bfd/
	* config.bfd: Remove K1_TCB_SIZE macro def for k1-*-*.
	* elfnn-k1.c (K1_TCB_SIZE): Remove.
	(tpoff_base): Remove use of K1_TCB_SIZE.

Differential Revision: D3563