Activate non-stop mode by default for kvx-gdb
Summary:
By default, gdb starts with non-stop mode disabled. As Kalray gdb
exposes all the hardware processors from the beginning, even the
ones which are not started yet, it can work only in the non-stop
mode. The attach-mppa command enables the non-stop mode before using
the target remote command to connect to the gdbserver communicating
with MPPA. This commit enables the non-stop mode by default for
kvx-gdb, so that the target remote command works.

	*gdb/kvx-tdep.c (kvx_gdbarch_init): Enable the non-stop mode.

Test Plan: GDB testsuite

Reviewers: O12 gdb!

Differential Revision: https://phab.kalray.eu/D4861