kvx: gas: rename XOR* mnemonics to EOR*
TCA mnemonics all start with an X, XOR was the only exception.
This renaming makes it so that only TCA mnemonics start by an X.
On KV3-1 and KV3-2, we install synthetics so that XOR still works.
However, the KV4-1 will not expose a XOR mnemonic.

include/ChangeLog:

2023-11-20  Paul Iannetta  <piannetta@kalrayinc.com>

        * opcode/kvx.h: Regenerate.

opcodes/ChangeLog:

2023-11-20  Paul Iannetta  <piannetta@kalrayinc.com>

        * kvx-opc.c: Regenerate.

gas/ChangeLog:

2023-11-20  Paul Iannetta  <piannetta@kalrayinc.com>

        * config/kvx-parse.h: Regenerate.